Removed relaxed YAML, only using processedYAML, warning badge will be determined based on # of passed tests
