+++ validator/constants.py
@@ -1,4 +1,4 @@
-RELAXED_SWAGGER = "relaxed-ga4gh-tool-discovery.yaml"
 SWAGGER = "processed-ga4gh-tool-discovery.yaml"
 GENERATED_SWAGGER = 'GA4GH.swagger.json'
 PREPROCESSED_SWAGGER = 'ga4gh-tool-discovery.yaml'
+EXPECTED_PASSING_TESTS = 9
